+++ terraform/dependencies/main.tf
@@ -3,10 +3,10 @@ provider "aws" {
 }
 
 terraform {
-  required_version = "~> 0.13.0"
+  required_version = "~> 1.0"
 
   required_providers {
-    aws = "3.22.0"
+    aws = "3.56"
   }
 
   backend "s3" {