+++ CHANGELOG.md
@@ -1,5 +1,8 @@
 # CHANGELOG
 
+## 2021-10-12
+* Updated terraform code to be 1.0 compliant
+
 ## 2021-07-26
 * Added `config/server.json`
 * Added `nodejs` to platform.yaml